Combined 'device' and 'api_key' options
Removed the "device" value option and combined it with the "api_keys" option. Now if the user wants to send Pushbullet pushes to a specific device, they should specify it in the King Phisher server configuration file by pre-pending it to their API key, and separating it with a colon. If only and API key is supplied, the default behaviour of pushing to all devices is still observed.

Example:
```
  plugins:
    pushbullet_notifications:
      api_keys: DEVICE:API_KEY
      identifier: 'King Phisher - Dev'
      mask: false
```

This allows for multiple API keys, separated by commas, while keeping the ability to push to all or specific devices, depending on each user's preference.